# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Leadership Engagement Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /people:
    get:
      operationId: listPeople
      tags: [people]
      summary: List all people
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [executive, secondary_leader, staff]
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of people
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Person"
    post:
      operationId: createPerson
      tags: [people]
      summary: Create a new person
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
  /people/{id}:
    get:
      operationId: getPerson
      tags: [people]
      summary: Get a person by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Person found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "404":
          description: Not found
    patch:
      operationId: updatePerson
      tags: [people]
      summary: Update a person
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "404":
          description: Not found
    delete:
      operationId: deletePerson
      tags: [people]
      summary: Delete a person
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /people/{id}/engagement:
    get:
      operationId: getPersonEngagement
      tags: [people]
      summary: Get engagement history for a person
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Engagement history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonEngagement"
        "404":
          description: Not found

  # ─── Events ───────────────────────────────────────────────────────────────
...
  /settings/{key}:
    patch:
      operationId: updateSetting
      tags: [settings]
      summary: Update a single setting value
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingUpdate"
      responses:
        "200":
          description: Updated setting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Setting"
        "404":
          description: Unknown key

  # ─── Audit Log ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Person:
      type: object
      required: [id, name, email, role, homeCity, homeState, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        title:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        role:
          type: string
          enum: [executive, secondary_leader, staff]
        homeCity:
          type: string
        homeState:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PersonInput:
      type: object
      required: [name, email, role, homeCity, homeState]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        title:
          type: string
        department:
          type: string
        role:
          type: string
          enum: [executive, secondary_leader, staff]
        homeCity:
          type: string
          minLength: 1
        homeState:
          type: string
          minLength: 1
        notes:
          type: string
    PersonUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        title:
          type: string
        department:
          type: string
        role:
          type: string
          enum: [executive, secondary_leader, staff]
        homeCity:
          type: string
        homeState:
          type: string
        notes:
          type: string
    PersonEngagement:
      type: object
      required: [person, invitations, virtualMeetings, daysSinceLastTouchpoint]
      properties:
        person:
          $ref: "#/components/schemas/Person"
        invitations:
          type: array
          items:
            $ref: "#/components/schemas/Invitation"
        virtualMeetings:
          type: array
          items:
            $ref: "#/components/schemas/VirtualMeeting"
        daysSinceLastTouchpoint:
          type: ["integer", "null"]
        totalInPersonAttended:
          type: integer
        totalVirtualCompleted:
          type: integer

    # ─── Event ──────────────────────────────────────────────────────────────
    Event:
      type: object
      required: [id, name, location, city, state, startDate, eventType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: string
        city:
          type: string
        state:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        eventType:
          type: string
          enum: [summit, conference, marketing, leadership, regional, other]
        leaderCount:
          type: integer
        inviteeCount:
          type: integer
        attendeeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    Invitation:
      type: object
      required: [id, eventId, personId, status, createdAt]
      properties:
        id:
          type: integer
        eventId:
          type: integer
        personId:
          type: integer
        status:
          type: string
          enum: [invited, attended, no_show, declined]
        notes:
          type: ["string", "null"]
        graphEventId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        person:
          $ref: "#/components/schemas/Person"
        event:
          $ref: "#/components/schemas/Event"
...
    VirtualMeeting:
      type: object
      required: [id, title, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        scheduledDate:
          type: ["string", "null"]
          format: date
        status:
          type: string
          enum: [suggested, scheduled, completed, cancelled]
        notes:
          type: ["string", "null"]
        hostId:
          type: ["integer", "null"]
        host:
          $ref: "#/components/schemas/Person"
        participantCount:
          type: integer
        teamsJoinUrl:
          type: ["string", "null"]
          description: "Microsoft Teams meeting join URL, populated when the meeting is scheduled via Graph API."
        graphMeetingId:
          type: ["string", "null"]
          description: "Microsoft Graph online meeting ID for cancellation."
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Setting:
      type: object
      required: [key, value, updatedAt]
      properties:
        key:
          type: string
        value:
          type: string
        updatedAt:
          type: string
          format: date-time
    SettingUpdate:
      type: object
      required: [value]
      properties:
        value:
          type: string

    # ─── Audit Log ───────────────────────────────────────────────────────────
...