# lib/api-zod/src/generated/types/setting.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Leadership Engagement Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Setting {
  key: string;
  value: string;
  updatedAt: Date;
}